emma suggestions
